+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.12.0-dev
 
+- [Add skip amount in `ModifyActions::skip`][105]
 - [Add multiple actions with a tuple][104]
     - Removes `add_many` as `add` can now be used instead
 - [Add cleanup of actions for despawned agents][101]
@@ -10,6 +11,7 @@
     - Replaces all unwraps with logging
     - Adds `on_remove_hook` and `on_remove_trigger` for both `CurrentAction` and `ActionQueue`
 
+[105]: https://github.com/hikikones/bevy-sequential-actions/pull/105
 [104]: https://github.com/hikikones/bevy-sequential-actions/pull/104
 [101]: https://github.com/hikikones/bevy-sequential-actions/pull/101